feat : Write Component 분리
