+++ 신동현/src/pages/index.tsx
@@ -4,8 +4,10 @@ import SignUpDetail from "./SignUpDetail";
 import Mypage from "./Mypage";
 import KakaoLoginLoading from "./KakaoLoginLoading";
 import Detail from "./Detail/Detail";
-import Write from "./Write";
+import Write from "./Write/Write";
+import WriteForm from "./Write/WriteForm";
+import Update from "./Update";
 import Root from "./Root";
 import NotFound from "./Not Found";
 
-export { Home, SignUp, SignUpDetail, Mypage, KakaoLoginLoading, Detail, Write, Root, NotFound };
+export { Home, SignUp, SignUpDetail, Mypage, KakaoLoginLoading, Detail, Write, WriteForm, Update, Root, NotFound };
